requirements.txt
flask
tensorflow>=2.10.0
numpy
scikit-learn
faiss-cpu
sentence-transformers
transformers
datasets
tqdm